backend/requirement.txt
torch
torchvision
torchaudio
opencv-python
tqdm
matplotlib
pillow
google-generativeai
fastapi
uvicorn
python-multipart
pandas
pyyaml
seaborn
scipy